Project is now pip installable with `pip install .`
This allows it to work as a `language: python` script for pre-commit.
It uses a pyproject.toml with Flit as the build tool.
Make sure you have a recent version of pip (21.3.1).

The project's file structure has been updated for use as a python package.

Removed leading 'v' from version scheme.